fix bug when first enter the page must hover on pins then the background image load
